DS4F gen-quality: mHC + bf16-KV + tokenizer/gen-mode + long-ctx decode bench
End-to-end greedy generation on real DeepSeek-V4-Flash weights now produces
coherent code (recursive quicksort), gated on DS4F_MHC=1 (hyper-connections,
hc_mult=4) — without mHC the residual collapses to a single stream and output
is garbage (TF next-token accuracy 0/23 -> 16/23 = 69.6% with mHC). The
argmax-exact-vs-batched check does NOT catch this; DS4F_TF_CHECK is the cheap
reference-free correctness gate.

- tools/ds4f_tokenizer.py: pure-Python byte-level BPE (encode/decode, stdlib only)
- ds4f_ep_runner.c: gen-mode (DS4F_PROMPT_IDS/DS4F_GEN_OUT/DS4F_MAX_NEW, greedy
  argmax feedback, eos=1) + DS4F_TF_CHECK teacher-forcing gate
- run_ds4f_gen_11n.sh: gen wrapper (defaults DS4F_MHC=1)
- bf16 KV cache (f32 -> uint16): halves the long-ctx KV footprint; bf16 not fp16
  (kv latents carry massive-activation dims that overflow fp16 -> Inf -> NaN)
- fix s_tb2_sel/s_idx_score heap overflow: index_topk=512 > max_pos corrupted the
  compressor ring-state -> NaNs; size to max(max_pos, index_topk)
- remove session-added cmpprobe/CMPSTEP debug scaffolding (DS4F_DEBUG-gated)
- long-ctx decode tok/s (mHC + bf16 KV): 2.99 @ctx10240, 2.68 @ctx16384 (below
  the 10-15 target; tb2prep index-scan 38-43% + attn 27% bound). ctx=32768 OOMs
  but is weight-footprint-bound (27.5GB dense, +6GB from FP8_BF16=1), NOT KV:
  bf16 KV alone does not raise the ceiling. ds4f.md Step 2e documents the gen
  repro, the decode table, and the 256k levers (FP8-dense / int8-KV / int8-SVE
  index_score).
- ds4f.md OPS #6 + run_ds4f_longctx_11n.sh: a sig=9 OOM-kill degrades the remote
  plexec/PMIx daemon (every later launch dies pre-load, rc=255 load=0/11);
  crash detector now flags this instead of reporting crash_hits=0.

+++ common/ds4f.h
@@ -191,8 +191,14 @@ typedef struct {
     float *hc_attn_fn, *hc_ffn_fn;    /* [24,16384] */
     float *hc_attn_base, *hc_ffn_base;  /* [24] */
     float *hc_attn_scale, *hc_ffn_scale; /* [3] */
-    /* per-layer KV cache: [max_pos, kv_lora] F32 (latent, 1 kv head) */
-    float *kv_cache;
+    /* per-layer KV cache: [max_pos, kv_lora] BF16 (latent, 1 kv head).
+     * 2-byte storage halves the long-context KV footprint (the decode memory
+     * dominator) vs f32. BF16 (not FP16): the model is natively bfloat16 and the
+     * kv latents carry massive-activation dims (~1e4..1e5) that overflow FP16's
+     * 65504 ceiling -> Inf -> NaN; BF16 shares f32's exponent range and the
+     * model's 8-bit-mantissa reference precision (faithful, not lossy). Codec:
+     * ds4f_bf16f (bf16->f32 lsl) / ds4f_f32bf (f32->bf16 round) in ds4f_impl.h. */
+    uint16_t *kv_cache;
     /* ---- Tier-B2 (DS4F_TIERB2; off-arena, only on ratio!=0 layers) ----
      * The stateful compressor/indexer long-range compressed-KV term that Tier-B1
      * omits. The matvec weights are stored bf16 (sources are bf16/FP8-e4m3, both
